pyproject.toml
[project]
name = "redis-mcp-server"
version = "0.2.0"
description = "Redis MCP Server - Model Context Protocol server for Redis"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Redis", email = "oss@redis.com"}
]
keywords = ["redis", "mcp", "model-context-protocol", "ai", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "mcp[cli]>=1.9.4",
    "redis>=6.0.0",
    "python-dotenv>=0.9.9",
    "numpy>=2.2.4",
    "click>=8.0.0",
    "uvicorn>=0.23.0",
    "starlette>=0.27.0",
]

[project.scripts]
redis-mcp-server = "src.main:cli"

[project.urls]
Homepage = "https://github.com/redis/mcp-redis"
Repository = "https://github.com/redis/mcp-redis"
Issues = "https://github.com/redis/mcp-redis/issues"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
